Don't guess where the branch island regions are, calculate it.
This fixes a problem where the code ended up thinking that a branch island
was not needed, when it was.

Example data:

__text section size: 0x4b90850, with 7 branch island regions.

Need to branch from 0x6A774 to 0x4B77250, thumb instruction.

branch island regions are at:
1:   0x0de0860
2:   0x14bd890
3:   0x21810e0
4:   0x2bcbe60
5:   0x3996a50
6:   0x476d680
7:   0x4b90850

Since target address is after branch island region #6, branch islands should be inserted into regions 1 through 6.

Previous code would assume a constant 14MB between each branch region, which would result in the following region addresses:

1:   0x0E00000
2:   0x1C00000
3:   0x2A00000
4:   0x3800000
5:   0x4600000
6:   0x5400000
7:   0x6200000

And the code would not insert a branch island into region #6, because it would think the target address is before region 6.

The final result would be that the branch from the island in region 5 to the 0x4B77250 would not fit into the thumb2 branch
instruction, overflowing it so that it turned negative, and branch into random memory.